Adjust memory map
Make use of the 0x8000-0x9000 memory area for code and data. This is
the low memory of the stack of the kernel and should be a good
starting point and it keeps all code and data into one single area
which makes it easier to avoid it in application code.