Implement single step machine instructions
On simpler targets this needs to be done by inserting a temporary
breakpoint. As I could not see anything in the protocol for this, I
decided to borrow the step command and change its semantics slightly
depending on the target. The optional address given is used taken as
the stop point for the next instruction to be exectued, not the actual
address to be executed.