Implement reset target command
This will clear local breakpoints and this command requires extended
mode so we say we support that.

Also the 'kill' command is not expected to send any reply, so we now
omit making any response other than a packet ack. Previously it sent
an empty reply, but that is not how it should be.