68000: Start in supervisor mode
The debugee starts in supervisor mode the first time as it inherits
the initial registers from the monitor program itself. This gives
flexibility in that the user program can install its own exception and
interrupt handling. However, if the user program drops down to user
mode, we need to ensure that if the program restarts it does not
inherit the last settings, but rather will restart in supervisor mode.