Change to step protocol
The step command expects the next PC address for setting a temporary
breakpoint. This is non-standard in the GDB server protocol, but I see
no other way to provide a proper single step while keeping the monitor
small and simple. Normally the current PC is passed over which allows
it to be altered. This commit restores that behavior in that the
current PC is passed over and the next PC is a second argument to
step. This makes it follow the protocol specification better. It also
means that it is a breaking change and you need to match the debugger
front-end properly to the monitor code.