Implement X command
This existed before but had several issues and was not used. Now the
Calypsi host debugger will use it. The main problem with it is that
we are potentially running on small targets with limited memory, so we
cannot rely on the usual buffer. Besides, binary data in the gdbserver
protocol works somewhat different than ASCII encoded packets.